+++ libs/ngx-mime/src/lib/viewer/viewer.component.ts
@@ -31,11 +31,10 @@ import { MimeViewerIntl } from '../core/intl';
 import { MimeResizeService } from '../core/mime-resize-service/mime-resize.service';
 import { MimeViewerConfig } from '../core/mime-viewer-config';
 import { ModeService } from '../core/mode-service/mode.service';
+import { RecognizedTextMode, ViewerMode } from '../core/models';
 import { Manifest } from '../core/models/manifest';
 import { ModeChanges } from '../core/models/modeChanges';
-import { RecognizedTextMode } from '../core/models/recognized-text-mode';
 import { ViewerLayout } from '../core/models/viewer-layout';
-import { ViewerMode } from '../core/models/viewer-mode';
 import { ViewerOptions } from '../core/models/viewer-options';
 import { ViewerState } from '../core/models/viewerState';
 import { StyleService } from '../core/style-service/style.service';
